geo_test/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.4
django-background-tasks==1.2.0
django-compat==1.0.15
django-rest-knox==3.6.0
djangorestframework==3.9.2
idna==2.8
Pillow==5.4.1
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
